chore: Refactoring everything for new relational classificator
