+++ .gitignore
@@ -13,10 +13,10 @@
 /prueba_py.fodg
 /utils/model_positions_copy.py
 /AI_PROJECT/output
-/AI_PROJECT/dataset/utils/__pycache__
-/AI_PROJECT/model/__pycache__
-/AI_PROJECT/dataset/__pycache__
-/AI_PROJECT/dataset/utils/transform/__pycache__
+/AI_PROJECT/datasets/utils/__pycache__
+/AI_PROJECT/models/__pycache__
+/AI_PROJECT/datasets/__pycache__
+/AI_PROJECT/datasets/utils/transform/__pycache__
 /AI_PROJECT/relator/__pycache__
 /AI_PROJECT/test_detect_and_relate.py
 /pred_boxes.pkl
