+++ AI_PROJECT/models/detector_functions.py
@@ -1,11 +1,11 @@
 import torch
 import torch.utils
 import torch.utils.data
-from dataset.dataset import visualize_images
-from dataset.utils.transformations import collate_function
+from datasets.dataset import visualize_images
+from datasets.utils.transformations import collate_function
 from torchmetrics.detection import MeanAveragePrecision
 import os
-from model.utils import (get_cuda_device, Averager, nms_on_output_dictionary,
+from models.utils import (get_cuda_device, Averager, nms_on_output_dictionary,
                          nms_filter_boxes, evaluate_predictions, test_transform)
 
 
@@ -110,13 +110,12 @@ def get_inference_and_metrics(weights_file, model, data_loader, num_classes, iou
             for i, output in enumerate(outputs):
                 gt_boxes = targets[i]['boxes'].cpu().numpy()
                 gt_labels = targets[i]['labels'].cpu().numpy()
-                
                 pred_boxes = output['boxes'].cpu().numpy()
                 pred_scores = output['scores'].cpu().numpy()
                 pred_labels = output['labels'].cpu().numpy()
 
-                tp, fp, fn = evaluate_predictions(gt_boxes, gt_labels, pred_boxes, pred_labels, pred_scores, iou_threshold)
-                
+                tp, fp = evaluate_predictions(gt_boxes, gt_labels, pred_boxes, pred_labels, iou_threshold)
+                #tp, fp, fn = evaluate_predictions(gt_boxes, gt_labels, pred_boxes, pred_labels, iou_threshold)
                 for class_id in range(1, num_classes+1):
                     y_true[class_id].extend([1] * tp[class_id] + [0] * fp[class_id])
                     y_scores[class_id].extend(pred_scores[pred_labels == class_id])
