+++ AI_PROJECT/perform_metrics.py
@@ -1,7 +1,7 @@
-from dataset.load_dataset import get_torch_dataloader
-from model.existent_model import inference_test, get_inference_and_metrics
-from model.utils import create_PRC
-from model.model_definition import model_defintion
+from datasets.detector_dataset import get_torch_dataloader
+from models.detector_functions import inference_test, get_inference_and_metrics
+from models.utils import create_PRC
+from models.detector_definition import model_defintion
 
 if __name__ == "__main__":
